automation/leads_monthly_samples/requirements.txt
jaydebeapi
pandas
pyyaml